// src/types/auth.ts
export enum ProviderOnboardingStep {
  PERSONAL_DETAILS = 'PERSONAL_DETAILS',
  CLINIC_AND_VISIT = 'CLINIC_AND_VISIT',
  INSURANCE_PROVIDERS = 'INSURANCE_PROVIDERS',
  WORK_EXPERIENCE = 'WORK_EXPERIENCE',
  EDUCATION = 'EDUCATION',
  CERTIFICATIONS = 'CERTIFICATIONS',
  SKILLS_AND_SPECIALIZATIONS = 'SKILLS_AND_SPECIALIZATIONS',
  SUBMIT_FOR_APPROVAL = 'SUBMIT_FOR_APPROVAL',
}